[Refactor] TourAPI DTO 사용 체계 조정
`#` TourAPI 내 모든 dto는 req/res 로 관리됨
`+` 응답 res를 인터페이스로 통합
`#` 변경된 패키지 위치 반영
