--- openapi/src/main/java/fc/be/openapi/tourapi/dto/bone/FacilityDTO.java
@@ -1,16 +1,0 @@
-package fc.be.openapi.tourapi.dto.bone;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.experimental.SuperBuilder;
-
-@SuperBuilder
-@NoArgsConstructor
-@Getter
-public class FacilityDTO extends PlaceDTO {
-    private String infoCenter;
-    private String usefee;
-    private String usetime;
-    private String restdate;
-    private String parking;
-}
