--- openapi/src/main/java/fc/be/openapi/tourapi/dto/bone/FestivalDTO.java
@@ -1,19 +1,0 @@
-package fc.be.openapi.tourapi.dto.bone;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.experimental.SuperBuilder;
-
-@SuperBuilder
-@NoArgsConstructor
-@Getter
-public class FestivalDTO extends PlaceDTO {
-    private String sponsor;
-    private String sponsorTel;
-    private String startDate;
-    private String endDate;
-    private String playtime;
-    private String eventPlace;
-    private String homepage;
-    private String usetime;
-}
