--- openapi/src/main/java/fc/be/openapi/tourapi/dto/form/common/Body.java
@@ -1,9 +1,0 @@
-package fc.be.openapi.tourapi.dto.form.common;
-
-public record Body<T>(
-        Items<T> items,
-        int numOfRows,
-        int pageNo,
-        int totalCount
-) {
-}
