--- openapi/src/main/java/fc/be/openapi/tourapi/dto/form/common/Header.java
@@ -1,7 +1,0 @@
-package fc.be.openapi.tourapi.dto.form.common;
-
-public record Header(
-        String resultCode,
-        String resultMsg
-) {
-}
