--- openapi/src/main/java/fc/be/openapi/tourapi/dto/form/common/Items.java
@@ -1,8 +1,0 @@
-package fc.be.openapi.tourapi.dto.form.common;
-
-import java.util.List;
-
-public record Items<T>(
-        List<T> item
-) {
-}
