--- openapi/src/main/java/fc/be/openapi/tourapi/dto/form/common/Response.java
@@ -1,7 +1,0 @@
-package fc.be.openapi.tourapi.dto.form.common;
-
-public record Response<T>(
-        Header header,
-        Body<T> body
-) {
-}
