+++ openapi/src/main/java/fc/be/openapi/tourapi/dto/response/bone/AccommodationDTO.java
@@ -1,0 +1,16 @@
+package fc.be.openapi.tourapi.dto.response.bone;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
+
+@SuperBuilder
+@NoArgsConstructor
+@Getter
+public class AccommodationDTO extends PlaceDTO {
+    private String infoCenter;
+    private String checkIn;
+    private String checkOut;
+    private String reservationUrl;
+    private String parking;
+}
