+++ openapi/src/main/java/fc/be/openapi/tourapi/dto/response/bone/FacilityDTO.java
@@ -1,0 +1,16 @@
+package fc.be.openapi.tourapi.dto.response.bone;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
+
+@SuperBuilder
+@NoArgsConstructor
+@Getter
+public class FacilityDTO extends PlaceDTO {
+    private String infoCenter;
+    private String usefee;
+    private String usetime;
+    private String restdate;
+    private String parking;
+}
