+++ openapi/src/main/java/fc/be/openapi/tourapi/dto/response/bone/LeportsDTO.java
@@ -1,0 +1,17 @@
+package fc.be.openapi.tourapi.dto.response.bone;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
+
+@SuperBuilder
+@NoArgsConstructor
+@Getter
+public class LeportsDTO extends PlaceDTO {
+    private String infoCenter;
+    private String openPeriod;
+    private String restDate;
+    private String useTime;
+    private String useFee;
+    private String parking;
+}
