+++ openapi/src/main/java/fc/be/openapi/tourapi/dto/response/bone/RestaurantDTO.java
@@ -1,0 +1,16 @@
+package fc.be.openapi.tourapi.dto.response.bone;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
+
+@SuperBuilder
+@NoArgsConstructor
+@Getter
+public class RestaurantDTO extends PlaceDTO {
+    private String firstMenu;
+    private String openTime;
+    private String restDate;
+    private String packing;
+    private String parking;
+}
