+++ openapi/src/main/java/fc/be/openapi/tourapi/dto/response/bone/ShopDTO.java
@@ -1,0 +1,15 @@
+package fc.be.openapi.tourapi.dto.response.bone;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
+
+@SuperBuilder
+@NoArgsConstructor
+@Getter
+public class ShopDTO extends PlaceDTO {
+    private String infoCenter;
+    private String restDate;
+    private String openTime;
+    private String parking;
+}
