+++ openapi/src/main/java/fc/be/openapi/tourapi/dto/response/form/TourAPIResponse.java
@@ -1,0 +1,10 @@
+package fc.be.openapi.tourapi.dto.response.form;
+
+import java.util.List;
+
+public interface TourAPIResponse {
+
+    boolean isItemEmpty();
+
+    <T> List<T> getItems();
+}
