+++ openapi/src/main/java/fc/be/openapi/tourapi/dto/response/form/common/Body.java
@@ -1,0 +1,9 @@
+package fc.be.openapi.tourapi.dto.response.form.common;
+
+public record Body<T>(
+        Items<T> items,
+        int numOfRows,
+        int pageNo,
+        int totalCount
+) {
+}
