+++ openapi/src/main/java/fc/be/openapi/tourapi/dto/response/form/common/Header.java
@@ -1,0 +1,7 @@
+package fc.be.openapi.tourapi.dto.response.form.common;
+
+public record Header(
+        String resultCode,
+        String resultMsg
+) {
+}
