+++ openapi/src/main/java/fc/be/openapi/tourapi/dto/response/form/common/Items.java
@@ -1,0 +1,8 @@
+package fc.be.openapi.tourapi.dto.response.form.common;
+
+import java.util.List;
+
+public record Items<T>(
+        List<T> item
+) {
+}
