+++ openapi/src/main/java/fc/be/openapi/tourapi/dto/response/form/common/Response.java
@@ -1,0 +1,7 @@
+package fc.be.openapi.tourapi.dto.response.form.common;
+
+public record Response<T>(
+        Header header,
+        Body<T> body
+) {
+}
